dashboard refactor + fix integrations tab spacing
