+++ dashboard/src/main/home/integrations/SlackIntegrationList.tsx
@@ -105,7 +105,6 @@ const Placeholder = styled.div`
   font-size: 13px;
   font-family: "Work Sans", sans-serif;
   justify-content: center;
-  margin-top: 40px;
   color: #aaaabb;
   border-radius: 5px;
   background: #26292e;
@@ -119,7 +118,6 @@ const Label = styled.div`
 `;
 
 const StyledIntegrationList = styled.div`
-  margin-top: 20px;
   margin-bottom: 80px;
 `;
